+++ test/buildComment.test.ts
@@ -109,11 +109,14 @@ describe('buildComment', () => {
             
             <details>
             
-            | Benchmark suite | Current: testCommitIdCurrent | Previous: testCommitIdPrevious | Ratio |
+            Previous: testCommitIdPrevious
+            Current: testCommitIdCurrent
+            
+            | Benchmark suite | Current | Previous | Ratio |
             |-|-|-|-|
-            | \`TestBench<1>\` | \`0\` testUnit | \`0\` testUnit | \`1\` |
-            | \`TestBench<2>\` | \`1\` testUnit | \`0\` testUnit | \`+∞\` |
-            | \`TestBench<3>\` | \`-1\` testUnit | \`0\` testUnit | \`-∞\` |
+            | \`TestBench<1>\` | \`0 testUnit\` | \`0 testUnit\` | \`1\` |
+            | \`TestBench<2>\` | \`1 testUnit\` | \`0 testUnit\` | \`+∞\` |
+            | \`TestBench<3>\` | \`-1 testUnit\` | \`0 testUnit\` | \`-∞\` |
             
             </details>
             
@@ -195,11 +198,14 @@ describe('buildComment', () => {
             
             <details>
             
-            | Benchmark suite | Current: testCommitIdCurrent | Previous: testCommitIdPrevious | Ratio |
+            Previous: testCommitIdPrevious
+            Current: testCommitIdCurrent
+            
+            | Benchmark suite | Current | Previous | Ratio |
             |-|-|-|-|
-            | \`TestBench<1>\` | \`0\` testUnit | \`0\` testUnit | \`1\` |
-            | \`TestBench<2>\` | \`0\` testUnit | \`1\` testUnit | \`+∞\` |
-            | \`TestBench<3>\` | \`0\` testUnit | \`-1\` testUnit | \`-∞\` |
+            | \`TestBench<1>\` | \`0 testUnit\` | \`0 testUnit\` | \`1\` |
+            | \`TestBench<2>\` | \`0 testUnit\` | \`1 testUnit\` | \`+∞\` |
+            | \`TestBench<3>\` | \`0 testUnit\` | \`-1 testUnit\` | \`-∞\` |
             
             </details>